test/requirements.txt
PySide6
python-pptx
google-generativeai
PyMuPDF
Pillow
python-dotenv
qtawesome